requirements.dev.txt
appdirs==1.4.4
appnope==0.1.0; sys_platform == "darwin" or platform_system == "Darwin"
aspy.yaml==1.3.0
astroid==2.4.1
atomicwrites==1.4.0; sys_platform == "win32"
attrs==19.3.0
autopep8==1.5.2
backcall==0.1.0
bandit==1.6.2
black==19.10b0
bleach==3.1.5
certifi==2020.4.5.1
cffi==1.14.0; sys_platform == "linux"
cfgv==3.1.0
chardet==3.0.4
click==7.1.2
codecov==2.0.22
colorama==0.4.3; sys_platform == "win32" or platform_system == "Windows"
coverage==5.1
cryptography==2.9.2; sys_platform == "linux"
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
docutils==0.16
entrypoints==0.3
filelock==3.0.12
flake8==3.8.1
flake8-black==0.1.1
gitdb==4.0.5
gitpython==3.1.2
identify==1.4.15
idna==2.9
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
jeepney==0.4.3; sys_platform == "linux"
jinja2==2.11.2
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==1.2.15
jupyterlab-server==1.1.3
keyring==21.2.1
lazy-object-proxy==1.4.3
markupsafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.2.0
multidict==4.7.5; python_version >= "3.6"
mypy==0.761
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.6
nodeenv==1.3.5
notebook==6.0.3
packaging==20.3
pandocfilters==1.4.2
parso==0.7.0
pathspec==0.8.0
pbr==5.4.5
pexpect==4.8.0; sys_platform != "win32"
pickleshare==0.7.5
pkginfo==1.5.0.1
pluggy==0.13.1
pre-commit==1.21.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0; sys_platform != "win32" or os_name != "nt"
py==1.8.1
pycodestyle==2.6.0
pycparser==2.20; sys_platform == "linux"
pydocstyle==5.0.2
pyflakes==2.2.0
pygments==2.6.1
pylint==2.5.2
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.2
pytest-cov==2.8.1
pytest-vcr==1.0.2
python-dateutil==2.8.1
python-dotenv==0.10.5
pywin32==227; sys_platform == "win32"
pywin32-ctypes==0.2.0; sys_platform == "win32"
pywinpty==0.5.7; os_name == "nt"
pyyaml==5.3.1
pyzmq==19.0.1
readme-renderer==26.0
regex==2020.5.13
requests==2.23.0
requests-toolbelt==0.9.1
secretstorage==3.1.2; sys_platform == "linux"
send2trash==1.5.0
six==1.14.0
smmap==3.0.4
snowballstemmer==2.0.0
stevedore==1.32.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
tox==3.15.0
tqdm==4.46.0
traitlets==4.3.3
twine==3.1.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
vcrpy==4.0.2
virtualenv==20.0.20
wcwidth==0.1.9
webencodings==0.5.1
wrapt==1.12.1
yarl==1.4.2; python_version >= "3.6"
